+++ services/api-server/src/simcore_service_api_server/api/routes/files.py
@@ -30,7 +30,7 @@ from starlette.datastructures import URL
 from starlette.responses import RedirectResponse
 
 from ..._meta import API_VTAG
-from ...exceptions.services_utils import DEFAULT_BACKEND_SERVICE_STATUS_CODES
+from ...exceptions.service_errors_utils import DEFAULT_BACKEND_SERVICE_STATUS_CODES
 from ...models.pagination import Page, PaginationParams
 from ...models.schemas.errors import ErrorGet
 from ...models.schemas.files import (
