+++ services/api-server/src/simcore_service_api_server/api/routes/solvers.py
@@ -9,7 +9,7 @@ from models_library.api_schemas_api_server.pricing_plans import ServicePricingPl
 from pydantic import ValidationError
 from pydantic.errors import PydanticValueError
 
-from ...exceptions.services_utils import DEFAULT_BACKEND_SERVICE_STATUS_CODES
+from ...exceptions.service_errors_utils import DEFAULT_BACKEND_SERVICE_STATUS_CODES
 from ...models.basic_types import VersionStr
 from ...models.pagination import OnePage, Page, PaginationParams
 from ...models.schemas.errors import ErrorGet
