+++ services/api-server/src/simcore_service_api_server/api/routes/solvers_jobs_getters.py
@@ -21,8 +21,8 @@ from servicelib.fastapi.requests_decorators import cancel_on_disconnect
 from servicelib.logging_utils import log_context
 from starlette.background import BackgroundTask
 
-from ...exceptions.custom import InsufficientCreditsError, MissingWalletError
-from ...exceptions.services_utils import DEFAULT_BACKEND_SERVICE_STATUS_CODES
+from ...exceptions.custom_errors import InsufficientCreditsError, MissingWalletError
+from ...exceptions.service_errors_utils import DEFAULT_BACKEND_SERVICE_STATUS_CODES
 from ...models.basic_types import LogStreamingResponse, VersionStr
 from ...models.pagination import Page, PaginationParams
 from ...models.schemas.errors import ErrorGet
