+++ services/api-server/src/simcore_service_api_server/exceptions/handlers/_custom_errors.py
@@ -1,7 +1,11 @@
 from fastapi import Request, status
 from starlette.responses import JSONResponse
 
-from .custom import CustomBaseError, InsufficientCreditsError, MissingWalletError
+from ..custom_errors import (
+    CustomBaseError,
+    InsufficientCreditsError,
+    MissingWalletError,
+)
 
 
 async def custom_error_handler(_: Request, exc: CustomBaseError):
