+++ services/api-server/src/simcore_service_api_server/services/catalog.py
@@ -12,7 +12,7 @@ from models_library.services import ServiceDockerData, ServiceType
 from pydantic import Extra, ValidationError, parse_obj_as, parse_raw_as
 from settings_library.catalog import CatalogSettings
 
-from ..exceptions.services_utils import service_exception_mapper
+from ..exceptions.service_errors_utils import service_exception_mapper
 from ..models.basic_types import VersionStr
 from ..models.schemas.solvers import LATEST_VERSION, Solver, SolverKeyId, SolverPort
 from ..utils.client_base import BaseServiceClientApi, setup_client_instance
