+++ services/api-server/src/simcore_service_api_server/services/director_v2.py
@@ -13,7 +13,7 @@ from starlette import status
 
 from ..core.settings import DirectorV2Settings
 from ..db.repositories.groups_extra_properties import GroupsExtraPropertiesRepository
-from ..exceptions.services_utils import service_exception_mapper
+from ..exceptions.service_errors_utils import service_exception_mapper
 from ..models.schemas.jobs import PercentageInt
 from ..utils.client_base import BaseServiceClientApi, setup_client_instance
