+++ services/api-server/src/simcore_service_api_server/services/log_streaming.py
@@ -10,7 +10,7 @@ from pydantic import NonNegativeInt
 from servicelib.logging_utils import log_catch
 from servicelib.rabbitmq import RabbitMQClient
 
-from ..exceptions.log_streaming import (
+from ..exceptions.log_streaming_errors import (
     LogStreamerNotRegisteredError,
     LogStreamerRegistionConflictError,
 )
