+++ services/api-server/src/simcore_service_api_server/services/storage.py
@@ -17,7 +17,7 @@ from pydantic import AnyUrl, PositiveInt
 from starlette.datastructures import URL
 
 from ..core.settings import StorageSettings
-from ..exceptions.services_utils import service_exception_mapper
+from ..exceptions.service_errors_utils import service_exception_mapper
 from ..models.schemas.files import File
 from ..utils.client_base import BaseServiceClientApi, setup_client_instance
