+++ services/api-server/tests/unit/test_errors_services_exception_handling.py
@@ -1,7 +1,7 @@
 import pytest
 from fastapi import HTTPException, status
 from httpx import HTTPStatusError, Request, Response
-from simcore_service_api_server.exceptions.services_utils import (
+from simcore_service_api_server.exceptions.service_errors_utils import (
     service_exception_mapper,
 )
